redis_celery/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# Data Processing
pandas==2.2.3
numpy==1.26.4

# OpenSearch & Vector DB
opensearch-py==2.4.2
qdrant-client==1.7.0
redis



# Utilities
python-dotenv==1.0.1
tqdm==4.66.5
pydantic>=2.0.0

anthropic
celery[redis]

